+++ src/App.vue
@@ -16,6 +16,7 @@ export default {
   },
   mounted () {
     this.$zircle.config({
+      debug: true,
       style: {
         theme: 'blue'
       }
@@ -26,5 +27,5 @@ export default {
 </script>
 <style>
   @import url('https://fonts.googleapis.com/css?family=Source+Sans+Pro:400,700');
-  @import url('https://use.fontawesome.com/releases/v5.1.0/css/all.css');
+  @import url('https://use.fontawesome.com/releases/v5.3.1/css/all.css');
 </style>